Phs site map eslint (#74)
* fixed colors and deleted the ts file for category queries

* supabase fix

* supabase types added

* fixed eslint and prettier things for all pages pertaining to site maps
