+++ src/components/userComponents/SiteMap/MapInteractionHandler.tsx
@@ -1,11 +1,11 @@
 import { useMap } from 'react-leaflet';
-import React, { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { LatLngExpression } from 'leaflet';
 /**
  * A React component that recenters the map to a given coordinate. This component does not render any visual content.
- * @param {Object} props - Component props.
- * @param {L.LatLng} props.center - The center to which the map should fly. Must include 'lat' and 'lng' properties.
- * @returns {null} Nothing is rendered by this component.
+ * @param props - Component props.
+ * @param props.center - The center to which the map should fly. Must include 'lat' and 'lng' properties.
+ * @returns Nothing is rendered by this component.
  */
 function RecenterMap({ center }: { center: LatLngExpression }) {
   const map = useMap();
