+++ src/supabase/category/queries.tsx
@@ -1,51 +1,30 @@
-'use client';
-
-import { CategoryRow } from '../../types/types';
 import supabase from '../client';
-
-// eslint-disable-next-line jsdoc/require-returns
-/**
- *
- * @param category_in
- */
-// eslint-disable-next-line import/prefer-default-export, camelcase
-// export async function get_category_color(category_in: string) {
-//     const { data, error } = await supabase.rpc('get_category_color', { category_in });
-//     if (error) {
-//       // Log the error message for debugging
-//       console.error(`An error occurred trying to generate color: ${error.message}`);
-//       // Throw a new error with the message for further handling
-//       throw new Error(`An error occurred trying to generate color: ${error.message}`);
-//     }
-//     return data;
-// }
+import { CategoryRow } from '../../types/types';
 // Assume this function is in `supabase/category/queries.js`
 /**
- *
- * @param category - The category for which you want to fetch the color
- * @returns The color for the category
+ * @param id - Each tour or exhibit has its unique identifier
+ * @returns color for category, else null
  */
 // eslint-disable-next-line import/prefer-default-export
-export async function getCategoryColor1(category: string) {
+export async function getCategoryColor1(id: string | number) {
   try {
     const { data, error } = await supabase
-      .from('categories') // Adjust according to your actual table name
-      .select('color')
-      .eq('category', category); // Use this if you're sure there's only one entry per category
+      .from('categories')
+      .select('color_hex')
+      .eq('id', id);
 
     if (error) {
-      throw new Error(error.message);
+      return null; // Return null on query error
     }
 
-    // Assuming 'color' is the column you want and each category has a unique color.
-    // 'data' would be the object containing 'color', so we return data.color.
-    // Check if 'data' exists to avoid accessing property 'color' of null.
-    console.log(`Color for ${category}:`, data[0].color);
-    return data ? data[0].color : null;
+    // Check if data array is not empty
+    if (data && data.length > 0 && data[0].color_hex) {
+      return data[0].color_hex;
+    }
+
+    return null; // Return null if no matching data found
   } catch (error) {
-    console.error(`Error fetching color for category ${category}:`, error);
-    // Return null or a default color to handle errors gracefully
-    return null;
+    return null; // Return null on unexpected error
   }
 }
